+++ lib/catechism.dart
@@ -2,7 +2,7 @@
 library;
 
 export 'src/application.dart';
-export 'src/configuration.dart';
-export 'src/models/info_screen_data.dart';
-export 'src/models/question.dart';
-export 'src/models/question_reference.dart';
+export 'src/features/configuration/domain/configuration.dart';
+export 'src/features/custom_pages/domain/custom_page_data.dart';
+export 'src/features/questions/domain/question.dart';
+export 'src/features/questions/domain/question_reference.dart';