+++ lib/src/features/custom_pages/data/custom_page_provider.dart
@@ -1,0 +1,4 @@
+import 'package:catechism/catechism.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+Provider<List<CustomPageData>>? customPageProvider;